typing_speed_app/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
arabic-reshaper==3.0.0
asn1crypto==1.5.1
bcrypt==4.0.1
Brotli==1.1.0
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
cryptography==44.0.3
cssselect2==0.8.0
defusedxml==0.7.1
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
fonttools==4.58.0
greenlet==3.2.2
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.4.0
MarkupSafe==3.0.2
oscrypto==1.3.0
Panda3D==1.10.15
passlib==1.7.4
pillow==11.2.1
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pydyf==0.11.0
pyHanko==0.27.1
pyhanko-certvalidator==0.26.8
pypdf==5.5.0
pyphen==0.17.2
python-bidi==0.6.6
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
qrcode==8.2
reportlab==4.4.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlmodel==0.0.24
starlette==0.46.2
svglib==1.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.4.0
uvicorn==0.34.2
weasyprint==65.1
webencodings==0.5.1
xhtml2pdf==0.2.17
zopfli==0.2.3.post1
